style(backend): split browsing in new files
