+++ backend/matcha/routes/__init__.py
@@ -1,6 +1,7 @@
 from flask import Blueprint
 
 import matcha.routes.users
+import matcha.routes.browsing
 import matcha.routes.auth
 import matcha.routes.images
 import matcha.routes.interests
@@ -10,6 +11,7 @@ import matcha.routes.like
 
 def init_routes(app):
     app.register_blueprint(users.bp)
+    app.register_blueprint(browsing.bp)
     app.register_blueprint(auth.bp)
     app.register_blueprint(images.bp)
     app.register_blueprint(interests.bp)
