+++ backend/matcha/services/users.py
@@ -23,7 +23,6 @@ from matcha.db.db import (
     db_set_user_profile_data,
     db_get_user_per_id,
     db_get_user_per_username,
-    db_browsing_gender_sexualorientation,
     db_get_url_profile,
 )
 
@@ -123,53 +122,6 @@ def services_users(id_user, request):
             ),
             200,
         )
-
-
-def services_browsing(id_user):
-    user_db = db_get_user_per_id(id_user)
-    gender = user_db[4]
-    sexual_orientation = user_db[5]
-
-    search = {gender: None, sexual_orientation: None}
-
-    if gender == "m" and sexual_orientation == "e":
-        search["gender"] = "f"
-        search["sexual_orientation"] = "e"
-    elif gender == "m" and sexual_orientation == "o":
-        search["gender"] = "m"
-        search["sexual_orientation"] = "o"
-    elif gender == "f" and sexual_orientation == "e":
-        search["gender"] = "m"
-        search["sexual_orientation"] = "e"
-    elif gender == "f" and sexual_orientation == "o":
-        search["gender"] = "f"
-        search["sexual_orientation"] = "o"
-
-    db_browsing_users = db_browsing_gender_sexualorientation(id_user, search)
-
-    browsing_users = []
-    for user in db_browsing_users:
-        profile_picture = db_get_url_profile(user[0])
-
-        if "url" in profile_picture:
-            profile_url = url = profile_picture["url"]
-        elif "error" in profile_picture:
-            profile_url = url = profile_picture["error"]
-
-        browsing_users.append(
-            {
-                "username": user[1],
-                "firstname": user[2],
-                "lastname": user[3],
-                "gender": user[4],
-                "sexualPreference": user[5],
-                "age": user[6],
-                "fameRating": user[7],
-                "urlProfile": profile_url,
-            }
-        )
-
-    return jsonify(browsing_users), 200
 
 
 def _email_put(user_id, request):
